# services/api/openapi/watch-store-api.yaml
openapi: 3.0.3
info:
  title: Watch Store API
  version: 0.1.0
  description: REST API for the Watch Store e-commerce platform
servers:
  - url: http://localhost:8080
    description: Local development
paths:
...
  /api/v1/webhooks/stripe:
    post:
      tags: [Payments]
      summary: Stripe webhook receiver
      description: |
        Receives signed Stripe webhook events. Requires the raw request body and
        `Stripe-Signature` header when Stripe is enabled. Events are processed
        idempotently by Stripe event ID.

        Handled event types:
        - `payment_intent.succeeded` → order status `PAID`
        - `payment_intent.payment_failed` → order status `FAILED`
      parameters:
        - in: header
          name: Stripe-Signature
          required: true
          schema:
            type: string
          description: HMAC signature from Stripe (`whsec_...` secret)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              description: Raw Stripe Event JSON payload
      responses:
        "200":
          description: Event accepted (including duplicate event IDs)
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: received
        "400":
          description: Missing or invalid Stripe signature
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...